move read_inventory_file to fab/utils.py
