+++ fab/utils.py
@@ -256,3 +256,23 @@ def bower_command(command, production=True, config=None):
 
 def get_inventory(inventory_path):
     return Inventory(loader=DataLoader(), variable_manager=VariableManager(), host_list=inventory_path)
+
+
+def read_inventory_file(filename):
+    """
+    filename is a path to an ansible inventory file
+
+    returns a mapping of group names ("webworker", "proxy", etc.)
+    to lists of hostnames as listed in the inventory file.
+    ("Hostnames" can also be IP addresses.)
+    If the hostname in the file includes :<port>, that will be included here as well.
+
+    """
+    inventory = get_inventory(filename)
+    port_map = {host.name: inventory.get_vars(hostname=host.name).get('ansible_port')
+                for host in inventory.get_hosts()}
+    return {group: [
+        '{}:{}'.format(host, port_map[host])
+        if port_map[host] is not None else host
+        for host in hosts
+    ] for group, hosts in get_inventory(filename).get_group_dict().items()}
